requirements.txt
pandas
numpy
websocket-client
matplotlib
